Test_task_3/Parse_API/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlalchemy
requests